pipelines/feature_generation/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
boto3==1.9.198
botocore==1.12.253        # via boto3, s3transfer
click==7.0
colorama==0.3.9           # via halo, log-symbols
cursor==1.2.0             # via halo
docutils==0.15.2          # via botocore
halo==0.0.26
jmespath==0.9.4           # via boto3, botocore
log-symbols==0.0.13       # via halo
nmslib==1.8.1
numpy==1.17.3
pillow==7.0.0             # via torchvision
pybind11==2.4.3           # via nmslib
python-dateutil==2.8.1    # via botocore
s3transfer==0.2.1         # via boto3
six==1.12.0               # via halo, python-dateutil, torchvision
spinners==0.0.23          # via halo
termcolor==1.1.0          # via halo
torch==1.3.1
torchvision==0.4.2
tqdm==4.26.0
urllib3==1.25.8           # via botocore
